ref(cmd/draft): move init ensure steps to file
+ purely for readability of init
+ using this as an opportunity to add tests here
