+++ cmd/draft/testdata/plugins/echo/plugin.yaml
@@ -2,5 +2,6 @@ name: echo
 usage: "echo stuff"
 description: "This echos stuff"
 command: "echo hello"
+version: "1.0.0"
 hooks:
     install: "echo installing"
